+++ DbConnector.Core/DbConnector.cs
@@ -683,6 +683,22 @@ namespace DbConnector.Core
         #region Implementation
 
         /// <summary>
+        /// Gets the string used to open the connection.
+        /// </summary>
+        public string ConnectionString
+        {
+            get { return _jobSetting.ConnectionString; }
+        }
+
+        /// <summary>
+        /// Gets the type of <see cref="DbConnection"/> being used.
+        /// </summary>
+        public Type ConnectionType
+        {
+            get { return _jobSetting.DbConnectionType; }
+        }
+
+        /// <summary>
         ///  <para>Creates an <see cref="IDbJob{IEnumerable{T}}"/> able to execute a reader based on the <paramref name="onInit"/> action.</para>
         ///  <para>Valid <typeparamref name="T"/> types: <see cref="DataSet"/>, <see cref="DataTable"/>, <see cref="Dictionary{string,object}"/>, any .NET built-in type, or any struct or class with a parameterless constructor not assignable from <see cref="System.Collections.IEnumerable"/> (Note: only properties will be mapped).</para>
         ///  See also: